docs(example-next-js-pages): use fetch client
