+++ examples/with-next-js-pages/package.json
@@ -16,7 +16,8 @@
     "clsx": "^2.1.1",
     "next": "^15.1.7",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "zimic": "latest"
   },
   "devDependencies": {
     "@playwright/test": "^1.50.1",
@@ -28,7 +29,6 @@
     "dotenv-cli": "^8.0.0",
     "postcss": "^8.5.2",
     "tailwindcss": "^4.0.6",
-    "typescript": "^5.7.3",
-    "zimic": "latest"
+    "typescript": "^5.7.3"
   }
 }
